+++ services/hdbpp-ttl/hdbpp_ttl.py
@@ -38,6 +38,7 @@ this script will delete dta from.
 
 import os
 import time
+import datetime
 import sys
 import argparse
 import math
@@ -47,6 +48,7 @@ import logging
 import logging.handlers
 
 from datetime import timedelta
+from hdbpp_rest_report import put_dict_to_rest
 
 logger = logging.getLogger('hdbpp-ttl')
 
@@ -130,12 +132,13 @@ def load_config(path):
     return config
 
 
-def process_ttl(server_config, dryrun):
+def process_ttl(server_config, dryrun, processed_ttl=None):
     """
     This function opens a connection to the server and processes all ttl values in the
     att_conf table. What this means in practice is it selects all values from att_conf 
     with a valid ttl value, then runs a delete for that attributes data in its data
     table
+    The processed_ttl dict is filled with the attributes that are treated
     """
 
     # first attempt to open a connection to the database
@@ -183,8 +186,13 @@ def process_ttl(server_config, dryrun):
                         attr[2], math.ceil(int(attr[1]) / 24), attr[0]
                     )
                 )
-
-                logger.info("Deleted {} rows in table: {} for attribute: {}".format(cursor.rowcount, attr[2], attr[3]))
+                
+                deleted_rows = cursor.rowcount
+                
+                if processed_ttl is not None:
+                    processed_ttl[attr[3]] = {"ttl_rows_deleted":deleted_rows}
+                
+                logger.info("Deleted {} rows in table: {} for attribute: {}".format(deleted_rows, attr[2], attr[3]))
 
             else:
                 cursor.execute(
@@ -252,9 +260,17 @@ def run_command(args):
             for key, val in config.items():
                 start_time = time.monotonic()
                 logger.info("Processing ttl requests for configuration: {}".format(key))
-                process_ttl(val["connection"], args.dryrun)
+                processed_ttl = {}
+                process_ttl(val["connection"], args.dryrun, processed_ttl)
                 delete_time = timedelta(seconds=time.monotonic() - start_time)
                 logger.info("Processed ttl request for configuration {} in: {}".format(key, delete_time))
+                
+                ttl_report = {"attributes":processed_ttl,
+                        "ttl_delete_duration":delete_time,
+                        "ttl_last_timestamp":datetime.datetime.now()}
+                
+                if "rest_endpoint" in val:
+                    put_dict_to_rest(val["rest_endpoint"], ttl_report)
 
         else:
             return False
@@ -266,7 +282,7 @@ def main() -> bool:
     parser = argparse.ArgumentParser(description="HDB TimscaleDb ttl service script")
     parser.add_argument("-v", "--version", action="store_true", help="version information")
     parser.add_argument("-d", "--debug", action="store_true", help="debug output for development")
-    parser.add_argument("-c", "--config", default="/etc/hdb/reorder.conf", help="config file to use")
+    parser.add_argument("-c", "--config", default="/etc/hdb/hdbpp_ttl.conf", help="config file to use")
     parser.add_argument("--validate", action="store_true", help="validate the given config file")
     parser.add_argument("--syslog", action="store_true", help="send output to syslog")
     parser.add_argument("--dryrun", action="store_true", help="do not actually do the delete, just simulate it")